src/main/java/com/example/org/barrage/domain/Barrage.java
package com.example.org.barrage.domain;

import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.databind.ObjectMapper;

public class Barrage {
    private String author;
    private String content;
    private String avatar;

    public static Barrage format(Object json) {
        try {
            return new ObjectMapper().readValue((String) json, Barrage.class);
        } catch (JsonProcessingException e) {
            return null;
        }
    }

    public String toJson() {
        try {
            return new ObjectMapper().writeValueAsString(this);
        } catch (JsonProcessingException e) {
            return "";
        }
    }

    public String getAuthor() {
        return author;
    }

    public void setAuthor(String author) {
        this.author = author;
    }

    public String getContent() {
        return content;
    }

    public void setContent(String content) {
        this.content = content;
    }

    public String getAvatar() {
        return avatar;
    }

    public void setAvatar(String avatar) {
        this.avatar = avatar;
    }
}
